Use abbot-cli to manage the tweet skill
